Reminders are loaded from reminders.csv on startup
Stopped using rufus scheduler and managed reminding internally via a list.
Added AReminder data structure which represents a reminder.

Old reminders are ignored on startup.
The bot apologises if a reminder is sent late.
The reminders are saved only if a new one is added because we need to change the
whole file when we remove a reminder, and old reminders are ignored on startup
anyway.